refactor(lab/experiments): update import paths to lib
